feat: 뉴스 상세 페이지 관련 타입 정의 추가 (#185)
