+++ frontend/src/pages/PN/types/comment.ts
@@ -1,0 +1,7 @@
+export interface Comment {
+  id: string;
+  author: string;
+  content: string;
+  createdAt: string;
+  isMyComment: boolean;
+}
