+++ frontend/src/pages/PN/types/newsDetail.ts
@@ -1,0 +1,15 @@
+import type { TimelineCard } from './timelineCard'
+
+export interface NewsDetail {
+  id: string
+  title: string
+  image?: string
+  updatedAt: string
+  bookmarked: boolean
+  timeline: TimelineCard[]
+  statistics: {
+    positive: number
+    neutral: number
+    negative: number
+  }
+}
