+++ frontend/src/pages/PN/types/timelineCard.ts
@@ -1,0 +1,7 @@
+export interface TimelineCard {
+  title: string
+  content: string
+  source: string[]
+  startAt: string
+  endAt: string
+}
